Polish "Add support for Spring Cloud Sleuth"
This commit restores the existing configuration so that Spring Cloud
Sleuth is auto-configured to expose traces to Wavefront if available.
If not available, the existing raw Open Tracing support is
auto-configured as a fallback.

The sample also now uses Spring Cloud Sleuth as this is more in line
with what the Spring ecosystem expects.

See gh-22