Add Bill of Materials (#46)
This commit adds a BOM with the dependencies that this project should
manage. To make sure that the BOM is valid, the project has been
restructured to use the BOM internally.

As a result, the root POM is only used to invoke the reactor and does
not need to be published anymore. A new parent, inheriting from the bom
provides the required build configuration and external dependency
management.